+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# main
+  * Bump minimum `tch` version to `0.13.0`.
+
 # v0.13.1 - 2022-09-16
   * Tweak the behavior of `set_overlay` to preserve visibility of existing overlays.
   * Add `window.is_overlay_visible()`.