+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Update to `wgpu` version `0.17`.
   * Expose `winit` support for `WindowTouchpadMagnifyEvent` and `WindowTouchpadRotateEvent` on supported platforms.
   * Bump minimum `tch` version to `0.13.0`.
+  * Bump supported `image` version to `0.25`.
 
 # v0.13.1 - 2022-09-16
   * Tweak the behavior of `set_overlay` to preserve visibility of existing overlays.